sign over all applicable square sizes by default
